feat/create separate function srollToSection for scrolling to some section
